refactor(agent-extension): split account usage profile into dedicated file
Move the account usage probe profile types, loader, and validators out of
profiles.go (876 lines) into account_usage_profile.go so profiles.go stays
under the revive 800-line file-length limit.
